requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.5
django-cors-headers==3.7.0
django-rest==0.8.7
djangorestframework==3.12.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
